+++ src/common/Token.php
@@ -13,6 +13,8 @@ class Token extends BaseObject implements TreeNodeInterface
     private $content;
     /** @var bool Whether then token is hidden with respect to `Symbol` definition */
     private $isHidden = false;
+    /** @var bool Whether then token is inline with respect to `Grammar` inline tokens */
+    private $isInline = false;
     /** @var array|null Match data for the token given from `preg_match()` */
     private $match;
     /** @var integer|null Position of the token in the input text */
@@ -27,14 +29,24 @@ class Token extends BaseObject implements TreeNodeInterface
      * @param null $offset Position of the token in the input text
      * @param bool $isHidden [since 1.4.0] Whether then token is hidden with respect to `Symbol`
      * definition
+     * @param bool $isInline [since 1.5.0] Whether then token is inline with respect to `Grammar`
+     * inline tokens
      */
-    public function __construct($type, $content, $match = null, $offset = null, $isHidden = false)
+    public function __construct(
+        $type,
+        $content,
+        $match = null,
+        $offset = null,
+        $isHidden = false,
+        $isInline = false
+    )
     {
         $this->type = $type;
         $this->content = $content;
         $this->match = $match;
         $this->offset = $offset;
         $this->isHidden = $isHidden;
+        $this->isInline = $isInline;
     }
 
     /**
@@ -63,6 +75,16 @@ class Token extends BaseObject implements TreeNodeInterface
     public function isHidden()
     {
         return $this->isHidden;
+    }
+
+    /**
+     * Whether then token is inline with respect to `Grammar` inline tokens
+     * @return bool
+     * @since 1.5.0
+     */
+    public function isInline()
+    {
+        return $this->isInline;
     }
 
     /**