+++ tests/unit/common/TokenTest.php
@@ -6,7 +6,7 @@ use VovanVE\parser\tests\helpers\BaseTestCase;
 
 class TokenTest extends BaseTestCase
 {
-    public function testDumpAsString()
+    public function testBasic()
     {
         $type = 'foo';
         $content = 'bar';
@@ -28,10 +28,26 @@ class TokenTest extends BaseTestCase
             }
         }
 
+        return $token;
+    }
+
+    /**
+     * @param Token $token
+     * @depends testBasic
+     */
+    public function testChildrenMatch($token)
+    {
         $this->assertTrue($token->areChildrenMatch([]));
         $this->assertFalse($token->areChildrenMatch(['lorem']));
         $this->assertFalse($token->areChildrenMatch(['ipsum', 'dolor']));
+    }
 
+    /**
+     * @param Token $token
+     * @depends testBasic
+     */
+    public function testMake($token)
+    {
         $this->assertNull($token->made());
         $token->make(42);
         $this->assertEquals(42, $token->made());
@@ -39,15 +55,36 @@ class TokenTest extends BaseTestCase
         $o = new \stdClass;
         $token->make($o);
         $this->assertSame($o, $token->made());
+    }
 
+    /**
+     * @param Token $token
+     * @depends testBasic
+     */
+    public function testHidden($token)
+    {
         $this->assertFalse($token->isHidden(), 'token must not be hidden by default');
         $hidden = new Token('a', 'b', null, null, true);
         $this->assertTrue($hidden->isHidden(), 'hidden token');
     }
 
-    public function testNoChild()
+    /**
+     * @param Token $token
+     * @depends testBasic
+     */
+    public function testInline($token)
     {
-        $token = new Token('foo', 'bar');
+        $this->assertFalse($token->isInline(), 'token must not be inline by default');
+        $inline = new Token('*', '*', null, null, false, true);
+        $this->assertTrue($inline->isInline(), 'inline token');
+    }
+
+    /**
+     * @param Token $token
+     * @depends testBasic
+     */
+    public function testNoChild($token)
+    {
         $this->setExpectedException(\OutOfBoundsException::class);
         $token->getChild(0);
     }